add bpu

+++ difftest/interface.cpp
@@ -12,8 +12,8 @@ extern "C" void update_indi(svBit cpu_is_mmio, svBit cpu_valid, int rcsr_id){
     state.rcsr_id = rcsr_id;
 }
 
-extern "C" void update_pc(long long pc, int inst){
-    state.pc = (uint64_t)pc;
+extern "C" void update_pc(int pc, int inst){
+    state.pc = (uint32_t)pc;
     state.inst = (uint32_t)inst;
 }
 
@@ -25,8 +25,8 @@ extern "C" void update_priv(int priv){
     state.priv = priv;
 }
 
-extern "C" void update_excep(svBit intr, long long cause, long long pc){
+extern "C" void update_excep(svBit intr, long long cause, int pc){
     state.intr = intr;
     state.cause = cause;
-    state.excep_pc = pc;
+    state.excep_pc = (uint32_t)pc;
 }
